+++ modules/domino/src/DominoSampler.cpp
@@ -49,12 +49,10 @@ Assignments DominoSampler::do_get_sample_assignments(
   IMP_LOG_TERSE("Sampling with " << known_particles.size() << " particles as "
                                  << known_particles << std::endl);
   IMP_USAGE_CHECK(known_particles.size() > 0, "No particles to sample");
-  Pointer<RestraintSet> rs =
-      get_model()->get_root_restraint_set();
   ParticlesTemp pt(known_particles.begin(), known_particles.end());
 
   SubsetFilterTables sfts = get_subset_filter_tables_to_use(
-      RestraintsTemp(1, rs), get_particle_states_table());
+      get_restraints(), get_particle_states_table());
   IMP_IF_LOG(TERSE) {
     IMP_LOG_TERSE("Filtering with ");
     for (unsigned int i = 0; i < sfts.size(); ++i) {
@@ -166,12 +164,10 @@ void DominoSampler::load_vertex_assignments(unsigned int node_index,
   typedef boost::graph_traits<MergeTree>::adjacency_iterator NeighborIterator;
   SubsetMap subset_map = boost::get(boost::vertex_name, mt_);
 
-  Pointer<RestraintSet> rs =
-      get_model()->get_root_restraint_set();
   // ParticlesTemp known_particles=
   // get_particle_states_table()->get_particles();
   SubsetFilterTables sfts = get_subset_filter_tables_to_use(
-      RestraintsTemp(1, rs), get_particle_states_table());
+      get_restraints(), get_particle_states_table());
   IMP::PointerMember<AssignmentsTable> sst =
       DiscreteSampler::get_assignments_table_to_use(sfts, max_states);
   ListSubsetFilterTable *lsft = nullptr;
@@ -202,13 +198,11 @@ void DominoSampler::load_vertex_assignments(unsigned int node_index,
   typedef boost::graph_traits<MergeTree>::adjacency_iterator NeighborIterator;
   SubsetMap subset_map = boost::get(boost::vertex_name, mt_);
 
-  Pointer<RestraintSet> rs =
-      get_model()->get_root_restraint_set();
   // ParticlesTemp known_particles=
   // get_particle_states_table()->get_particles();
   // ParticlesTemp pt(known_particles.begin(), known_particles.end())
   SubsetFilterTables sfts = get_subset_filter_tables_to_use(
-      RestraintsTemp(1, rs), get_particle_states_table());
+      get_restraints(), get_particle_states_table());
   ListSubsetFilterTable *lsft = nullptr;
   if (csf_) {
     lsft = new ListSubsetFilterTable(get_particle_states_table());